test(framework): move tproxy tests config from test framework
This Config is and will be used only in transparent proxy tests,
so we can safely move it there. As now there is no conflict in
names, I renamev it from `TProxyConfig` to just `Config`.
